pyproject.toml
[project]
name = "hb-dsbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
]
[tool.ruff]
line-length = 120
